Replacing authenticated client from gdax-node to local executor class closes #25, closes #26, closes #27
