+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "gdax-flash-order",
-  "version": "0.2.4",
+  "version": "0.2.5",
   "description": "An algorithmic approach to getting market making orders filled on GDAX cryptocurrency exchange",
   "main": "index.js",
   "scripts": {
@@ -33,6 +33,8 @@
   },
   "dependencies": {
     "async": "^2.6.0",
+    "axios": "^0.18.0",
+    "crypto": "^1.0.1",
     "dotenv": "^6.0.0",
     "events": "^2.0.0",
     "gdax": "^0.6.0",
